Fix fuzz tests

+++ tests/fuzz/fuzz_d2atmastash.cc
@@ -3,7 +3,7 @@
 extern "C" int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
 	d2itemreader_init_default();
 	d2atmastash stash;
-	uint32_t bytesRead;
+	size_t bytesRead;
 	d2err err = d2atmastash_parse(data, size, &stash, &bytesRead);
 	if (err == D2ERR_OK)
 	{
