+++ tests/fuzz/fuzz_d2char.cc
@@ -3,7 +3,7 @@
 extern "C" int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
 	d2itemreader_init_default();
 	d2char character;
-	uint32_t bytesRead;
+	size_t bytesRead;
 	d2err err = d2char_parse(data, size, &character, &bytesRead);
 	if (err == D2ERR_OK)
 	{
